move config/auth to envs
