+++ src/app/(backend)/api/webhooks/clerk/route.ts
@@ -1,8 +1,8 @@
 import { NextResponse } from 'next/server';
 
-import { authEnv } from '@/config/auth';
 import { isServerMode } from '@/const/version';
 import { serverDB } from '@/database/server';
+import { authEnv } from '@/envs/auth';
 import { pino } from '@/libs/logger';
 import { UserService } from '@/server/services/user';
