+++ src/libs/next-auth/sso-providers/azure-ad.ts
@@ -1,6 +1,6 @@
 import AzureAD from 'next-auth/providers/azure-ad';
 
-import { authEnv } from '@/config/auth';
+import { authEnv } from '@/envs/auth';
 
 import { getMicrosoftEntraIdIssuer } from './microsoft-entra-id-helper';
 import { CommonProviderConfig } from './sso.config';
