+++ src/libs/next-auth/sso-providers/logto.ts
@@ -1,6 +1,6 @@
 import { OIDCConfig, OIDCUserConfig } from '@auth/core/providers';
 
-import { authEnv } from '@/config/auth';
+import { authEnv } from '@/envs/auth';
 
 import { CommonProviderConfig } from './sso.config';
