+++ src/libs/next-auth/sso-providers/microsoft-entra-id-helper.ts
@@ -1,4 +1,4 @@
-import { authEnv } from '@/config/auth';
+import { authEnv } from '@/envs/auth';
 
 function getTenantId() {
   return (
