+++ src/libs/next-auth/sso-providers/zitadel.ts
@@ -1,6 +1,6 @@
 import Zitadel from 'next-auth/providers/zitadel';
 
-import { authEnv } from '@/config/auth';
+import { authEnv } from '@/envs/auth';
 
 const provider = {
   id: 'zitadel',
